refactor: 질문 목록 조회 로직의 위치를 이동
기존 `/` 경로로 매핑된 서블릿의 질문 목록 조회 로직을 `/questions` 경로에 매핑된 서블릿으로 이동함.
`index.jsp`가 `/questions` 경로로 리다이렉트하도록 구현함.
