+++ src/main/webapp/index.jsp
@@ -1,0 +1,4 @@
+<%@ page contentType="text/html; charset=UTF-8" pageEncoding="utf-8" language="java" %>
+<%
+    response.sendRedirect(request.getContextPath() + "/questions");
+%>
